requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==23.2.0
chardet==4.0.0
diskcache==5.6.3
idna==3.6
Jinja2==3.1.3
llama_cpp_python==0.2.47
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
python-discord==1.7.3
python-dotenv==1.0.1
typing_extensions==4.9.0
yarl==1.9.4
